server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
anthropic==0.39.0
python-dotenv==1.0.1
